phase1_rss/requirements.txt
feedparser>=6.0.11
requests>=2.32.0
python-dotenv>=1.0.1
jinja2>=3.1.0

